Battery icon: force show pct in QS header and when charging/powersave
like aosp already wanted to do (QuickStatusBarHeader setForceShowPercent
was being overridden by the tuner observer added by custom battery
commits to BatteryMeterView).

This is also good for users that don't want the percentage in the
statusbar but still want to be able to check it in the QS panel.


PS 3: Fix percentage not showing in qs/keyguard after screen off / on
and force it only when hidden icon style or hidden percentage option
are enabled

Change-Id: I57cdd6c158af8440ed3372c7799875d34fd58f75